refactor(mongo_adapter): découpage par domaine (users/emails/mailboxes/calendar)
Splitte src/logic/mongo_adapter.rs (896 LOC) en sous-modules par domaine:
- mod.rs (324 LOC): struct + trait DatabaseInterface (délégations)
- users.rs (215), emails.rs (242), mailboxes.rs (127),
  mailboxes_user.rs (225), calendar.rs (97)

Chaque sous-module expose des méthodes inhérentes '*_impl' sur
MongoDatabaseAdapter ; le trait DatabaseInterface (dans mod.rs) délègue
via self.NAME_impl(...).await. Aucun changement de comportement.
